style(generate): format
